fix: 64-bit storage sizes — 8 GB card was displayed as 3.2 GB
size_t/uint32_t wrapped at 4 GB in the download page and in
pcap_fs_total/used(), which also feeds the >80%-full warning.

+++ src/main.cpp
@@ -319,8 +319,8 @@ void loop() {
             ui_pcap_update();
 
             // Storage warning when >80% full
-            uint32_t total = pcap_fs_total();
-            uint32_t used  = pcap_fs_used();
+            uint64_t total = pcap_fs_total();
+            uint64_t used  = pcap_fs_used();
             if (total > 0 && used > total * 4 / 5)
                 ui_pcap_show_storage_warning(true);
         }
